+++ src/main/java/np/edu/nast/payroll/Payroll/service/TaxSlabService.java
@@ -5,9 +5,14 @@ import java.util.List;
 
 public interface TaxSlabService {
 
+    // Use Integer to match the taxSlabId type in the TaxSlab entity
     TaxSlab create(TaxSlab slab);
-    TaxSlab update(Long id, TaxSlab slab);
-    void delete(Long id);
-    TaxSlab getById(Long id);
+
+    TaxSlab update(Integer id, TaxSlab slab);
+
+    void delete(Integer id);
+
+    TaxSlab getById(Integer id);
+
     List<TaxSlab> getAll();
 }
